change the way currencies are represented

+++ src/ledger/LedgerGateway.h
@@ -28,7 +28,7 @@ namespace stellar
 		// called by CLF
         virtual void recvDelta(CLFDeltaPtr delta, LedgerHeaderPtr header) = 0;
 
-        virtual int64_t getTxFee() = 0;
+        virtual int32_t getTxFee() = 0;
         virtual int64_t getLedgerNum() = 0;
 	};
 }
